# a2/openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Task API",
      "version": "1.0.0",
      "description": "A simple in-memory to-do list API built with Express and TypeScript."
    },
    "servers": [{ "url": "http://localhost:3000" }],
    "paths": {
      "/": {
        "get": {
          "summary": "API information",
          "responses": {
            "200": { "description": "Basic info about the API" }
          }
        }
      },
      "/health": {
        "get": {
          "summary": "Health check",
          "responses": {
            "200": {
              "description": "Server is running",
              "content": {
                "application/json": {
                  "example": { "status": "ok" }
                }
              }
            }
          }
        }
      },
      "/tasks": {
        "get": {
          "summary": "Get all tasks",
          "responses": {
            "200": {
              "description": "List of tasks",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Task" }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a task",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateTask" },
                "example": { "title": "Buy milk" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Task created",
              "content": {
                "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
              }
            },
            "400": {
              "description": "Invalid input",
              "content": {
                "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
              }
            }
          }
        }
      },
      "/tasks/{id}": {
        "get": {
          "summary": "Get one task",
          "parameters": [
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
          ],
          "responses": {
            "200": {
              "description": "The task",
              "content": {
                "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
              }
            },
            "404": {
              "description": "Task not found",
              "content": {
                "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
              }
            }
          }
        },
        "put": {
          "summary": "Update a task",
          "parameters": [
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpdateTask" },
                "example": { "title": "Buy milk and bread", "done": true }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated task",
              "content": {
                "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
              }
            },
            "400": {
              "description": "Invalid input",
              "content": {
                "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
              }
            },
            "404": {
              "description": "Task not found",
              "content": {
                "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
              }
            }
          }
        },
        "delete": {
          "summary": "Delete a task",
          "parameters": [
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
          ],
          "responses": {
            "204": { "description": "Task deleted" },
            "404": {
              "description": "Task not found",
              "content": {
                "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Task": {
          "type": "object",
          "properties": {
            "id": { "type": "integer", "example": 1 },
            "title": { "type": "string", "example": "Learn Express" },
            "done": { "type": "boolean", "example": false }
          }
        },
        "CreateTask": {
          "type": "object",
          "required": ["title"],
          "properties": {
            "title": { "type": "string", "example": "Buy milk" }
          }
        },
        "UpdateTask": {
          "type": "object",
          "properties": {
            "title": { "type": "string", "example": "Buy milk and bread" },
            "done": { "type": "boolean", "example": true }
          }
        },
        "Error": {
          "type": "object",
          "properties": {
            "error": { "type": "string", "example": "Task 99 not found" }
          }
        }
      }
    }
  }